pyproject.toml
[project]
name = "cloud-storage-api"
version = "0.1.0"
description = "Cloud Storage API"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "fastapi>=0.128.0",
    "pydantic[email]>=2.12.5",
    "pydantic-settings>=2.12.0",
    "sqlalchemy>=2.0.45",
    "uvicorn[standard]>=0.40.0",
    "passlib>=1.7.4",
    "pyjwt>=2.10.1",
    "redis>=7.1.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.10",
    "types-passlib>=1.7.7.20250602",
]

[tool.mypy]
python_version = "3.13"
strict = true

plugins = ["pydantic.mypy"]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[tool.ruff]
line-length = 88
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "migrations",
]
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"

addopts = [
    "--verbose",
    "--strict-markers",
    "--strict-config",
]
